\name{getDefaultFilters}
\alias{getDefaultFilters}
\title{Get default set of filters...}
\description{Get default set of filters}
\value{Default set of filters for jchoose.files}
\seealso{{\code{\link{jchoose.files}}, \code{\link{tkchoose.files}}, \code{\link{rchoose.files}}}}
\author{Alex Lisovich, Roger Day}
