\name{getDefaultFilters}
\alias{getDefaultFilters}
\title{Get default set of filters}
\usage{
  getDefaultFilters()
}
\value{
  Default set of filters for jchoose.files
}
\description{
  Get default set of filters
}
\author{
  Alex Lisovich, Roger Day
}
\seealso{
  {\code{\link{jchoose.files}},
  \code{\link{tkchoose.files}},
  \code{\link{rchoose.files}}}
}

