\name{rChoiceDialogs-package}
\title{rChoiceDialogs.}
\description{rChoiceDialogs.}
\details{\tabular{ll}{
Package: \tab rChoiceDialogs\cr
Type: \tab Package\cr
Version: \tab 1.0.3 \cr
Date: \tab 2012-07-27 \cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}}
\alias{rChoiceDialogs}
\docType{package}
\author{Alex Lisovich, Roger Day}
\keyword{package}
